fix: 무제한 쿠폰의 Lua 발급 호환성 복구
무제한 메타데이터와 시작 후 무발급 준비를 지원하고 기존 유한 Redis 상태 호환성을 유지한다. 벤치마크 프로필은 명령 범위로 제한한다.

+++ src/test/java/kr/kro/airbob/domain/coupon/CouponConcurrencyTest.java
@@ -241,7 +241,7 @@ class CouponConcurrencyTest {
 			return invocation.callRealMethod();
 		}).when(stockManager).prepare(
 			eq(transitioningCoupon.getId()),
-			eq((long)COUPON_LIMIT),
+			eq(COUPON_LIMIT),
 			anyLong(),
 			anyLong(),
 			eq(true),
